Fix clone method reference issue
The metatable mt needs the transfer function to be declared above it.
Cloning an ASource happens in the transfer function, and references the
local mt metatable... that's at that point in the code, undeclared.

One could have hoisted out the clone method elsewhere, but sometimes,
the simplest methods are the best of choices.