Only fill one internal buffer per iteration.
Implementation forced all OpenALSoft internal buffers to be filled, even
if we early returned once from the processing function due to the Source
not being a looping one, filling it with silence, but the next buffer
would be processed like that did not happen at all.

Fixes issue #3 .

Fix could have been made in multiple ways, this seemed the simplest, and
it shouldn't be worse in terms of performance either.